shipsenseai-azure-native/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.2
azure-search-documents==11.4.0
azure-storage-blob==12.20.0
blinker==1.8.2
camelot-py==0.11.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==7.0.0
cryptography==42.0.8
dataclasses-json==0.6.7
distro==1.9.0
elastic-transport==8.13.1
elasticsearch==8.14.0
et-xmlfile==1.1.0
etelemetry==0.3.1
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.4.1
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.11
langchain-experimental==0.0.62
langchain-text-splitters==0.2.2
langsmith==0.1.83
libclang==18.1.1
looseversion==1.3.0
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
openai==1.35.10
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.0
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pdfminer.six==20240706
pillow==10.4.0
protobuf==4.25.3
prov==2.0.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydot==2.0.0
Pygments==2.18.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
PyMySQL==1.1.1
pyodbc==5.1.0
pyparsing==3.1.2
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==6.3.2
regex==2024.5.15
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
setuptools==70.2.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
sympy==1.12.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
termcolor==2.4.0
tf_keras==2.16.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.4
traits==6.3.2
transformers==4.42.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
